+++ src/StateSpaceDynamics.jl
@@ -81,6 +81,6 @@ export info_update!
 export tview
 
 # Common functions
-export rand, smooth, fit!, loglikelihood, elbo, elbo!
+export rand, smooth, posterior, fit!, loglikelihood, elbo, elbo!
 
 end